PSP: Improved image viewer movement
Before the limit was arbitrary. Now we check the proper location on screen and move in increments relative to the size on screen.

svn-id: r53918

+++ backends/platform/psp/image_viewer.h
@@ -66,8 +66,8 @@ private:
 	void loadLastImage();
 	void setViewerButtons(bool active);
 	void setConstantRendererOptions();
-	void moveImageX(int val);
-	void moveImageY(int val);
+	void moveImageX(float val);
+	void moveImageY(float val);
 	bool load(int imageNum);
 	void unload();	
 	void runLoop(); // to get total pausing we have to do our own loop
